Added invalid.html and fixed bugs in delete and edit product functionality

+++ marketplace/myapp/views.py
@@ -21,6 +21,8 @@ def create_product(request):
     return render(request,'myapp/create_product.html',{'product_form':product_form})
 def product_edit(request,id):
     product=Product.objects.get(id=id)
+    if product.seller!=request.user:
+        return redirect('invalid')
     product_form=ProductForm(request.POST or None,request.FILES or None,instance=product)
     if request.method=='POST':
         if product_form.is_valid():
@@ -30,12 +32,14 @@ def product_edit(request,id):
     return render(request,'myapp/product_edit.html',{'product_form':product_form ,'product':product})
 def delete_product(request,id):
     product=Product.objects.get(id=id)
+    if product.seller!=request.user:
+        return redirect('invalid')
     if request.method=='POST':
         product.delete()
         return redirect('index')
     return render(request,'myapp/delete.html',{'product':product})
 def dashboard(request):
-    products=Product.objects.all()
+    products=Product.objects.filter(seller=request.user)
     return render(request,'myapp/dashboard.html',{'products':products})
 def register(request):
     if request.method=="POST":
@@ -49,3 +53,5 @@ def register(request):
 def user_logout(request):
     logout(request)
     return render(request,'myapp/logout.html')
+def invalid(request):
+    return render(request,'myapp/invalid.html')
